+++ Data/SBV/Core/Symbolic.hs
@@ -1398,8 +1398,7 @@ newUninterpreted st (nm, mbArgNames) t uiCode
                                                        , "  Name: " ++ nm
                                                        , "  Type: " ++ show t
                                                        , ""
-                                                       , "You should use these functions at least once the query part starts"
-                                                       , "and then use them in the query section as usual."
+                                                       , "You should explicitly register these functions by calling registerFunction before starting the query section."
                                                        ]
                                      pure True
                       UICgC c  -> -- No need to record the code in interactive mode: CodeGen doesn't use interactive